Fix frontend pre-commit and move unlocalized strings check to pre-commit hook (#7763)
